+++ ios/RCTPSPDFKit/RCTPSPDFKitView.m
@@ -30,6 +30,7 @@
 @property (nonatomic, nullable) UIViewController *topController;
 @property (nonatomic, strong) SessionStorage *sessionStorage;
 @property (nonatomic) BOOL isPropsSet;
+@property (nonatomic, strong) UITapGestureRecognizer *tapGestureRecognizer;
 
 @end
 
@@ -57,11 +58,6 @@
     [NSNotificationCenter.defaultCenter addObserver:self selector:@selector(spreadIndexDidChange:) name:PSPDFDocumentViewControllerSpreadIndexDidChangeNotification object:nil];
       
     [NSNotificationCenter.defaultCenter addObserver:self selector:@selector(documentDidFinishRendering) name:PSPDFDocumentViewControllerDidConfigureSpreadViewNotification object:nil];
-      
-    UITapGestureRecognizer *tapGestureRecognizer = [[UITapGestureRecognizer alloc] init];
-    [tapGestureRecognizer addTarget:self action:@selector(tapGestureRecognizerDidChangeState:)];
-    [_pdfController.interactions.allInteractions allowSimultaneousRecognitionWithGestureRecognizer:tapGestureRecognizer];
-    [_pdfController.view addGestureRecognizer:tapGestureRecognizer];
   }
   
   return self;
@@ -247,6 +243,21 @@
 }
 
 - (void)pdfViewController:(PSPDFViewController *)pdfController willBeginDisplayingPageView:(PSPDFPageView *)pageView forPageAtIndex:(NSInteger)pageIndex {
+    BOOL hasPageClickGesture = NO;
+    for (UIGestureRecognizer *gesture in pageView.gestureRecognizers) {
+        if ([gesture.name isEqualToString:@"reactNativePageClick"]) {
+            hasPageClickGesture = YES;
+            break;
+        }
+    }
+    if (hasPageClickGesture == NO) {
+        UITapGestureRecognizer *tapGestureRecognizer = [[UITapGestureRecognizer alloc] init];
+        tapGestureRecognizer.name = @"reactNativePageClick";
+        [tapGestureRecognizer addTarget:self action:@selector(tapGestureRecognizerDidChangeState:)];
+        [_pdfController.interactions.allInteractions allowSimultaneousRecognitionWithGestureRecognizer:tapGestureRecognizer];
+        [pageView addGestureRecognizer:tapGestureRecognizer];
+    }
+
   [self onStateChangedForPDFViewController:pdfController pageView:pageView pageAtIndex:pageIndex];
 }